requirements.txt
asgiref
certifi
crispy-bootstrap5
distlib
dj-database-url
dj-rest-auth
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
django-phone-field
django-solo
djangorestframework
filelock
gunicorn
Markdown
MouseInfo
numpy
opencv-contrib-python
packaging
pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
python-decouple
pytweening
setuptools
sqlparse
typing_extensions
tzdata
virtualenv
whitenoise
